style: Label 객체 구성 수정 및 객체명 변경
