+++ src/components/common/Label.tsx
@@ -1,4 +1,4 @@
-import { Hierarchy, labelType, Weight } from '@/styles/label';
+import { Hierarchy, labelStyle, Weight } from '@/styles/labelStyle';
 
 interface LabelProps {
   hierarchy: Hierarchy;
@@ -11,7 +11,7 @@ interface LabelProps {
 function Label({ weight, hierarchy, text, textColor, isRequired }: LabelProps) {
   return (
     <div
-      className={`${labelType.weight[weight][hierarchy].typo} ${labelType.weight[weight][hierarchy].lineHeight} gap-5xs flex`}
+      className={`${labelStyle.weight[weight].hierarchy[hierarchy].typo} ${labelStyle.weight[weight].hierarchy[hierarchy].lineHeight} gap-5xs flex`}
     >
       <span className={`${textColor} whitespace-nowrap`}>{text}</span>
       {isRequired && <span className='text-feedback-notification-dark'>*</span>}
